// src/app/models/procedureparam.ts
export class ProcedureParam {
    pCampo0: string = 'T';
    pValor0: string = '';
    pCampo1: string = 'T';
    pValor1: string = '';
    pCampo2: string = 'T';
    pValor2: string = '';
    pCampo3: string = 'T';
    pValor3: string = '';
    pCampo4: string = 'T';
    pValor4: string = '';
    pCampo5: string = 'T';
    pValor5: string = '';
    pCampo6: string = 'T';
    pValor6: string = '';
    pCampo7: string = 'T';
    pValor7: string = '';
    public inicia() {
      this.pCampo0 = 'T';
      this.pValor0 = '';
      this.pCampo1 = 'T';
      this.pValor1 = '';
      this.pCampo2 = 'T';
      this.pValor2 = '';
      this.pCampo3 = 'T';
      this.pValor3 = '';
      this.pCampo4 = 'T';
      this.pValor4 = '';
      this.pCampo5 = 'T';
      this.pValor5 = '';
      this.pCampo6 = 'T';
      this.pValor6 = '';
      this.pCampo7 = 'T';
      this.pValor7 = '';
    }
  }
  